refactor: 폴더 구조 최적화 및 블루버튼 수정
