+++ src/pages/main/lost/Lost.styles.ts
