+++ src/pages/main/lost/LostComplete.styles.ts
