+++ src/pages/main/lost/LostPost.styles.ts
