+++ src/pages/main/lost/LostPost.types.ts
