+++ src/pages/main/lost/LostUpload.styles.ts
